requirements-dev.txt
pre-commit==3.7.1
python-dotenv
